requirements.txt
Pillow
typing; python_version<"3.5"